extract CreateSet<T>() to a TableServices class
