+++ Runtime/Assist/ITableServices.cs
@@ -1,0 +1,9 @@
+﻿using System.Collections.Generic;
+
+namespace TechTalk.SpecFlow.Assist
+{
+    public interface ITableServices
+    {
+        IEnumerable<T> CreateSet<T>(Table table);
+    }
+}
