+++ Runtime/Assist/TableServices.cs
@@ -1,0 +1,34 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace TechTalk.SpecFlow.Assist
+{
+    public class TableServices : ITableServices
+    {
+        public static ITableServices Current
+        {
+            get
+            {
+                //TODO: Get it from current scenario context
+                return new TableServices();
+            }
+        }
+
+        public IEnumerable<T> CreateSet<T>(Table table)
+        {
+            var list = new List<T>();
+
+            var pivotTable = new PivotTable(table);
+            for (var index = 0; index < table.Rows.Count(); index++)
+            {
+                var instance = pivotTable.GetInstanceTable(index).CreateInstance<T>();
+                list.Add(instance);
+            }
+
+            return list;
+        }
+    }
+}
