+++ Runtime/TechTalk.SpecFlow.csproj
@@ -56,10 +56,12 @@
     </Compile>
     <Compile Include="Assist\EnumerableProjection.cs" />
     <Compile Include="Assist\FormattingTableDiffExceptionBuilder.cs" />
+    <Compile Include="Assist\ITableServices.cs" />
     <Compile Include="Assist\SafetyTableDiffExceptionBuilder.cs" />
     <Compile Include="Assist\SetComparer.cs" />
     <Compile Include="Assist\TableDifferenceResults.cs" />
     <Compile Include="Assist\TableDiffExceptionBuilder.cs" />
+    <Compile Include="Assist\TableServices.cs" />
     <Compile Include="Assist\TEHelpers.cs" />
     <Compile Include="Assist\InstanceComparisonExtensionMethods.cs" />
     <Compile Include="Assist\IValueComparer.cs" />
